Create Meeting API- table contents
