+++ src/main/java/com/app/katchup/Meeting/model/MeetingConstants.java
@@ -1,0 +1,17 @@
+package com.app.katchup.Meeting.model;
+
+public class MeetingConstants {
+
+    public static final String userNameNullMessage = "Username cannot be null";
+    public static final String passwordNullMessage = "Password cannot be null";
+    public static final String dateTimeNullMessage = "Date/Time cannot be null";
+    public static final String inviteListNullMessage = "To list cannot be null";
+    public static final String venueNullMessage = "Venue cannot be null";
+    public static final String responseNullMessage = "Response cannot be null";
+    public static final String durationNullMessage = "Duration cannot be null";
+    public static final String pollNullMessage = "Poll response cannot be null";
+    public static final int userNameMinLength = 5;
+    public static final int userNameMaxLength = 10;
+
+
+}
