+++ src/main/java/com/app/katchup/Meeting/model/Poll.java
@@ -1,0 +1,6 @@
+package com.app.katchup.Meeting.model;
+
+public enum Poll {
+    POLL,
+    NO_POLL,
+}
