--- src/main/java/com/app/katchup/MeetingResponse/model/Poll.java
@@ -1,6 +1,0 @@
-package com.app.katchup.MeetingResponse.model;
-
-public enum Poll {
-    POLL,
-    NOPOLL,
-}
