+++ src/main/java/com/app/katchup/Users/UserConstants.java
@@ -3,12 +3,6 @@ package com.app.katchup.Users;
 public class UserConstants {
     public static final String userNameNullMessage = "Username cannot be null";
     public static final String passwordNullMessage = "Password cannot be null";
-    public static final String dateTimeNullMessage = "Date/Time cannot be null";
-    public static final String toListNullMessage = "To list cannot be null";
-    public static final String venueNullMessage = "Venue cannot be null";
-    public static final String responseNullMessage = "Response cannot be null";
-    public static final String durationNullMessage = "Duration cannot be null";
-    public static final String pollNullMessage = "Poll response cannot be null";
     public static final int userNameMinLength = 5;
     public static final int userNameMaxLength = 10;
     static final int passwordMinLength = 8;
