+++ src/main/resources/application.properties
@@ -2,7 +2,6 @@ server.port=8080
 spring.data.mongodb.database=KatchUpDatabase
 spring.data.mongodb.port=27017
 spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration
-spring.data.mongodb.repositories.enabled=true
 spring.data.mongodb.host=localhost
 #logging.level.org.springframework.security= DEBUG
 logging.level.org.hibernate= DEBUG
